requirements.txt
pytest==7.4.1
coverage==7.3.2
vcrpy==5.1.0
requests==2.31.0